11/requirements.txt
openai
python-dotenv
pydub
tiktoken 